# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/half-register": {
      "post": {
        "operationId": "AuthController_register_half",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterHalfDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/full_register": {
      "post": {
        "operationId": "AuthController_register_full",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFullDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/complete_register": {
      "post": {
        "operationId": "AuthController_register_complete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCompleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/log_out": {
      "post": {
        "operationId": "AuthController_log_out",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogOutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset-request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password-reset-confirm": {
      "post": {
        "operationId": "AuthController_confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/user/Add_Review": {
      "post": {
        "operationId": "UserController_Add_Review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/Edit_Review": {
      "post": {
        "operationId": "UserController_Edit_Review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/Remove_Review": {
      "post": {
        "operationId": "UserController_Remove_Review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/Add_Cart_Item": {
      "post": {
        "operationId": "UserController_Add_Cart_Item",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/Remove_Cart_Item": {
      "post": {
        "operationId": "UserController_Remove_Cart_Item",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/increase_Quantity": {
      "post": {
        "operationId": "UserController_increase_Quantity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncreaseQuantityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/decrease_Quantity": {
      "post": {
        "operationId": "UserController_decrease_Quantity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecreaseQuantityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/products_search": {
      "get": {
        "operationId": "UserController_SearchProducts",
        "parameters": [
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "description": "Search keyword for product name or description",
            "schema": {
              "example": "laptop",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/ProductCategory": {
      "get": {
        "operationId": "UserController_FindProductsByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "query",
            "description": "Category ID or name to filter products",
            "schema": {
              "example": "electronics",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    }
  },
  "info": {
    "title": "Portfolio API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterHalfDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email to start partial registration"
          }
        },
        "required": [
          "email"
        ]
      },
      "RegisterFullDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email used for full registration step"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Verification code sent to email"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "RegisterCompleteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com",
            "description": "User email for final registration step"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "description": "Password for the new account"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email used for login"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "description": "User password used for login"
          }
        }
      },
      "LogOutDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email of the user to log out"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email used to send password reset token"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetConfirmDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email associated with the reset token"
          },
          "token": {
            "type": "string",
            "example": "abcdef123456",
            "description": "Password reset token sent to user email"
          },
          "newPassword": {
            "type": "string",
            "example": "NewStrongPassword1!",
            "description": "New password to set for the account"
          }
        },
        "required": [
          "email",
          "token",
          "newPassword"
        ]
      },
      "AddReviewDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email of the user submitting the review"
          },
          "comment": {
            "type": "string",
            "example": "Great product, highly recommended!",
            "description": "User comment on the product"
          },
          "productId": {
            "type": "string",
            "example": "8f2e9f13-7d34-4c5a-95b1-5e5b3b5c1c9a",
            "description": "UUID of the product being reviewed"
          },
          "rating": {
            "type": "number",
            "example": 4,
            "description": "Rating between 1 and 5"
          }
        },
        "required": [
          "email",
          "comment",
          "productId",
          "rating"
        ]
      },
      "EditReviewDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email of the user editing the review"
          },
          "reviewId": {
            "type": "string",
            "example": "a14b5c3e-8e9f-4c2d-923b-34b7f26a1b2c",
            "description": "UUID of the review to edit"
          },
          "comment": {
            "type": "string",
            "example": "Updated comment about the product",
            "description": "Updated review text"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "description": "Updated rating between 1 and 5"
          }
        },
        "required": [
          "email",
          "reviewId"
        ]
      },
      "RemoveReviewDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email of the user removing the review"
          },
          "reviewId": {
            "type": "string",
            "example": "a14b5c3e-8e9f-4c2d-923b-34b7f26a1b2c",
            "description": "UUID of the review to remove"
          }
        },
        "required": [
          "email",
          "reviewId"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "cartId": {
            "type": "string",
            "example": "7a2b8c4f-3c5d-4a6b-9e5f-12345abcde67",
            "description": "UUID of the user's cart"
          },
          "productId": {
            "type": "string",
            "example": "8f2e9f13-7d34-4c5a-95b1-5e5b3b5c1c9a",
            "description": "UUID of the product to add to the cart"
          }
        },
        "required": [
          "cartId",
          "productId"
        ]
      },
      "RemoveCartItemDto": {
        "type": "object",
        "properties": {
          "cartId": {
            "type": "string",
            "example": "7a2b8c4f-3c5d-4a6b-9e5f-12345abcde67",
            "description": "UUID of the user's cart"
          },
          "productId": {
            "type": "string",
            "example": "8f2e9f13-7d34-4c5a-95b1-5e5b3b5c1c9a",
            "description": "UUID of the product to remove from the cart"
          }
        },
        "required": [
          "cartId",
          "productId"
        ]
      },
      "IncreaseQuantityDto": {
        "type": "object",
        "properties": {
          "cartItemId": {
            "type": "string",
            "example": "3a5b9c8d-2e4f-4b6c-9a8f-34567f9e0b1a",
            "description": "UUID of the cart item to increase quantity"
          }
        },
        "required": [
          "cartItemId"
        ]
      },
      "DecreaseQuantityDto": {
        "type": "object",
        "properties": {
          "cartItemId": {
            "type": "string",
            "example": "3a5b9c8d-2e4f-4b6c-9a8f-34567f9e0b1a",
            "description": "UUID of the cart item to decrease quantity"
          }
        },
        "required": [
          "cartItemId"
        ]
      }
    }
  }
}